tests/test_content_filters.py
import unittest
from extensions.content_filters import BaseContentFilter

class TestBaseContentFilter(unittest.TestCase):
    def setUp(self):
        # Initialize the BaseContentFilter instance
        self.filter = BaseContentFilter()

    def test_filter_content(self):
        # Test the filter_content method of BaseContentFilter
        # Assuming it should return the content unchanged if no specific filter is applied
        test_content = "This is a sample content with <script>JavaScript content</script>"
        filtered_content = self.filter.filter_content(test_content)
        self.assertEqual(filtered_content, test_content, "The filter_content method should return the original content if no filters are applied.")

    def test_filter_not_implemented(self):
        # Test that the filter_content method raises a NotImplementedError when not overridden
        with self.assertRaises(NotImplementedError):
            self.filter.filter_content("Some content")

# Add more tests for other content filter classes when they are implemented
# For example:
# class TestAdvancedContentFilter(unittest.TestCase):
#     def setUp(self):
#         # Initialize the AdvancedContentFilter instance
#         self.filter = AdvancedContentFilter()
#
#     def test_advanced_filtering(self):
#         # Test the advanced filtering logic
#         test_content = "Some advanced content to filter"
#         expected_content = "Expected filtered content"
#         filtered_content = self.filter.filter_content(test_content)
#         self.assertEqual(filtered_content, expected_content, "The AdvancedContentFilter should filter content according to its logic.")

if __name__ == '__main__':
    unittest.main()
